disable pip publishing
After testing, (and realizing something I should have known at the start), pip cannot package non-python files such as .html or .js files. so when publishing and installing this as a package, none of those files are included, and thus, the application is not able to run. It only works on systems that have installed agvis through github. Therefore, pip is pretty much useless with AGVis, and conda would be a better place to handle such publishing tasks.